+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Add `download_report_document` helper method to Reports API (2021-06-30)
+
 ### [4.4.0] - 2025-05-30
 
 ### Added